+++ docs/hello-world/assets/js/main.1.221276c2.js
@@ -1,0 +1,2 @@
+console.log("\n>> Contact page");
+//# sourceMappingURL=main.1.221276c2.js.map