+++ docs/hello-world/assets/js/main.2.a50020ff.js
@@ -1,0 +1,2 @@
+console.log("\n>> Home page");
+//# sourceMappingURL=main.2.a50020ff.js.map