+++ docs/hello-world/assets/js/main.c4764d45.js
@@ -1,0 +1,2 @@
+console.log("\n>> About page");
+//# sourceMappingURL=main.c4764d45.js.map