+++ examples/hello-world/src/views/pages/home/main.js
@@ -1,6 +1,1 @@
-// include core module `prismjs` from installed npm package
-require('prismjs/prism.js');
-// include support for pug
-require('prismjs/components/prism-pug.min');
-
 console.log('\n>> Home page');